+++ chibicc-main/hashmap.c
@@ -14,8 +14,8 @@
 // Represents a deleted hash entry
 #define TOMBSTONE ((void *)-1)
 
-static uint64_t fnv_hash(char *s, int len) {
-  uint64_t hash = 0xcbf29ce484222325;
+static uint32_t fnv_hash(char *s, int len) {
+  uint32_t hash = 0xcbf29ce484222325;
   for (int i = 0; i < len; i++) {
     hash *= 0x100000001b3;
     hash ^= (unsigned char)s[i];
