+++ chibicc-main/tokenize.c
@@ -356,7 +356,7 @@ static bool convert_pp_int(Token *tok) {
     base = 8;
   }
 
-  int64_t val = strtoul(p, &p, base);
+  int32_t val = strtoul(p, &p, base);
 
   // Read U, L or LL suffixes.
   bool l = false;
